Rename refactored chrome_agent_client to chrome_agent and chrome_agent to chrome_file_server.
