+++ chrome_agent/js/extension/background.js
@@ -14,5 +14,5 @@
 
 __axiomExport__(this);
 
-// Create a chromeAgent and start it listening for a connection.
-var chromeAgent = new chrome_agent.ChromeAgent();
+// Create a ChromeFileServer and start it listening for a connection.
+var chromeFileServer = new chrome_agent.ChromeFileServer();
