+++ chrome_agent/manifest.json
@@ -16,7 +16,7 @@
       "js/lib/wash.amd.concat.js",
       "js/chrome_agent_client.js",
       "js/chrome_command.js",
-      "js/chrome_agent.js",
+      "js/chrome_file_server.js",
       "js/extension/background.js"
     ]
   },
